Refactored builder components
